fixing progress counter bug (#197)
* fixing progress counter bug; consolidating count printing logic to a single class;
* need to reset unflushed counter after flushing
* refactor to HashMap and change names for clearer code
* implementing spacing suggestion
* address review comments to consolidate the pritprogress function

---------

+++ src/main/java/com/datastax/cdm/properties/IPropertyHelper.java
@@ -29,5 +29,5 @@ public interface IPropertyHelper {
 
     boolean isSparkConfFullyLoaded();
 
-    boolean meetsMinimum(String valueName, Integer testValue, Integer minimumValue);
+    boolean meetsMinimum(String valueName, Long testValue, Long minimumValue);
 }
